feat(expo): now expo loads appropriate components
