+++ src/services/app-info.service.rn.ts
@@ -1,0 +1,12 @@
+import DeviceInfo from 'react-native-device-info';
+
+export class AppInfoService {
+
+  static getVersion = (): string => {
+    return DeviceInfo.getVersion();
+  };
+
+  static getBuildNumber = (): string => {
+    return DeviceInfo.getBuildNumber();
+  };
+}
