+++ src/services/app-reload.service.rn.ts
@@ -1,0 +1,8 @@
+import { NativeModules } from 'react-native';
+
+export class AppReloadService  {
+
+  static reload = (): void => {
+    NativeModules.DevMenu.reload();
+  };
+}
