+++ src/services/app-reload.service.ts
@@ -1,8 +1,8 @@
-import { NativeModules } from 'react-native';
+import Constants from 'expo-constants';
 
-export class AppReloadService  {
+import { AppReloadService as RNService } from './app-reload.service.rn';
+import { AppReloadService as ExpoService } from './app-reload.service.expo';
 
-  static reload = (): void => {
-    NativeModules.DevMenu.reload();
-  };
-}
+const Service = Constants.platform ? ExpoService : RNService;
+
+export const AppReloadService = Service;
